Both logging and unit tests use write_app_uart to redirect stdout to the uart.
